+++ test/int/test_iproto_options.py
@@ -49,27 +49,35 @@ def test_sql_vdbe_opcode_max_and_sql_motion_row_max_options(instance: Instance):
     instance.sql("SELECT * FROM (VALUES (1), (2)) OPTION (sql_vdbe_opcode_max = 100)")
 
 
-def test_invalid_sql_options(instance: Instance):
+def test_repeated_sql_options(instance: Instance):
+    instance.sql("""
+                 SELECT * FROM (VALUES (1), (2)) OPTION (
+                   sql_motion_row_max = 1,
+                   sql_motion_row_max = 10
+                 )""")
     with pytest.raises(
         TarantoolError,
-        match=r"sbroad: invalid query: option sql_motion_row_max specified more than once!",
+        match=r"unexpected number of values: Exceeded maximum number of rows \(1\) in virtual table: 2",
     ):
         instance.sql("""
                      SELECT * FROM (VALUES (1), (2)) OPTION (
-                       sql_motion_row_max = 3,
-                       sql_motion_row_max = 2,
-                       sql_vdbe_opcode_max = 1,
-                       sql_vdbe_opcode_max = 2
+                       sql_motion_row_max = 10,
+                       sql_motion_row_max = 1
                      )""")
 
+    instance.sql("""
+                 SELECT * FROM (VALUES (1), (2)) OPTION (
+                   sql_vdbe_opcode_max = 1,
+                   sql_vdbe_opcode_max = 100
+                 )""")
     with pytest.raises(
         TarantoolError,
-        match=r"sbroad: invalid query: option sql_vdbe_opcode_max specified more than once!",
+        match=r"Reached a limit on max executed vdbe opcodes. Limit: 1",
     ):
         instance.sql("""
                      SELECT * FROM (VALUES (1), (2)) OPTION (
-                       sql_vdbe_opcode_max = 1,
-                       sql_vdbe_opcode_max = 2
+                       sql_vdbe_opcode_max = 10,
+                       sql_vdbe_opcode_max = 1
                      )""")